merge from develop
* feat: replace all imports (#1)

* fix: import internal edwards25519 package to project (#2)

The following code depends on the internal package golang.org/x/crypto/ed25519/internal/edwards25519, so if you want to use it you need to know that this package was copied from the internal ed25519 package to make it available to your code. I’ve basically just copied necessary code snippets of edward25519 package needed into an existing project to make this work.